requirements.txt
# Core Web Framework
Flask==1.1.1
Werkzeug==0.16.0
Jinja2==2.11.3
itsdangerous==1.1.0
MarkupSafe==1.1.1

# Server
gevent

# Deep Learning
tensorflow==2.7.2
protobuf==3.20.3

# Image & Data Processing
numpy==1.22.0
pandas==1.0.1
Pillow==10.4.0
opencv-python==4.2.0.32
opencv-contrib-python==4.2.0.32
imutils==0.5.3
h5py==2.10.0

# ML Utilities
scikit-learn==0.22.1
scipy==1.4.1
numba==0.47.0
tqdm==4.42.1
